+++ src/node/neighborhood/transaction_response.go
@@ -4,5 +4,5 @@ type TransactionResponse struct {
 	Timestamp        int64
 	SenderAddress    string
 	RecipientAddress string
-	Value            float32
+	Value            uint64
 }
